+++ db/compaction/compaction.cc
@@ -242,9 +242,7 @@ Compaction::Compaction(
       trim_ts_(_trim_ts),
       is_trivial_move_(false),
       compaction_reason_(_compaction_reason),
-      notify_on_compaction_completion_(false),
-      blob_garbage_collection_policy_(_blob_garbage_collection_policy),
-      blob_garbage_collection_age_cutoff_(_blob_garbage_collection_age_cutoff) {
+      notify_on_compaction_completion_(false) {
   MarkFilesBeingCompacted(true);
   if (is_manual_compaction_) {
     compaction_reason_ = CompactionReason::kManualCompaction;
@@ -269,6 +267,26 @@ Compaction::Compaction(
   }
 
   GetBoundaryKeys(vstorage, inputs_, &smallest_user_key_, &largest_user_key_);
+
+  // selectively force-enable/force-disable blob garbage collection
+  if (_blob_garbage_collection_policy ==
+      BlobGarbageCollectionPolicy::kDisable) {
+    enable_blob_garbage_collection_ = false;
+  } else {
+    enable_blob_garbage_collection_ =
+        _blob_garbage_collection_policy ==
+            BlobGarbageCollectionPolicy::kForce ||
+        mutable_cf_options()->enable_blob_garbage_collection;
+  }
+
+  // selectively override blob age cutoff
+  if (_blob_garbage_collection_age_cutoff < 0 ||
+      _blob_garbage_collection_age_cutoff > 1) {
+    blob_garbage_collection_age_cutoff_ =
+        mutable_cf_options()->blob_garbage_collection_age_cutoff;
+  } else {
+    blob_garbage_collection_age_cutoff_ = _blob_garbage_collection_age_cutoff;
+  }
 }
 
 Compaction::~Compaction() {