+++ db/compaction/compaction.h
@@ -83,7 +83,7 @@ class Compaction {
              double score = -1, bool deletion_compaction = false,
              CompactionReason compaction_reason = CompactionReason::kUnknown,
              BlobGarbageCollectionPolicy blob_garbage_collection_policy =
-                 BlobGarbageCollectionPolicy::kUserDefault,
+                 BlobGarbageCollectionPolicy::kUseDefault,
              double blob_garbage_collection_age_cutoff = -1);
 
   // No copying allowed
@@ -310,19 +310,10 @@ class Compaction {
   uint32_t max_subcompactions() const { return max_subcompactions_; }
 
   bool enable_blob_garbage_collection() const {
-    if (blob_garbage_collection_policy_ ==
-        BlobGarbageCollectionPolicy::kDisable) {
-      return false;
-    }
-    return blob_garbage_collection_policy_ ==
-               BlobGarbageCollectionPolicy::kForce ||
-           mutable_cf_options()->enable_blob_garbage_collection;
+    return enable_blob_garbage_collection_;
   }
 
   double blob_garbage_collection_age_cutoff() const {
-    if (blob_garbage_collection_age_cutoff_ < 0) {
-      return mutable_cf_options()->blob_garbage_collection_age_cutoff;
-    }
     return blob_garbage_collection_age_cutoff_;
   }
 
@@ -434,12 +425,10 @@ class Compaction {
   // begin.
   bool notify_on_compaction_completion_;
 
-  // Force-enable/force-disable GC collection for blobs during compaction.
-  BlobGarbageCollectionPolicy blob_garbage_collection_policy_;
+  // Enable/disable GC collection for blobs during compaction.
+  bool enable_blob_garbage_collection_;
 
   // Blob garbage collection age cutoff.
-  // Note that it will not override the age cutoff in the DB options if it is
-  // a negative value.
   double blob_garbage_collection_age_cutoff_;
 };